+++ skyobject/CHANGELOG.md
@@ -3,14 +3,28 @@ changes
 
 ### TODO
 
-- [ ] add ability to remove all non-full root objects
 - [ ] move `node.Rootwalker` here
-- [ ] add tests for `(*Container).RangeFeed`
-- [ ] add tests for `(*Container).RootByHash`
-- [ ] add test for `(*Root).Index`
-- [ ] add test for `(*Root).Slice`
+- [ ] add test for `RemoveNonFullRoots`
+
 
 ### Done
+
+#### 10:00 28 June 2017
+
+- [x] add ability to remove all non-full root objects
+- [x] add test for `(*Root).Index`
+- [x] add test for `(*Root).Slice`
+- [x] add tests for `(*Container).RangeFeed`
+- [x] add tests for `(*Container).RootByHash`
+
+and
+
+- add test for `(*Root).Len`
+
+Introducing `(*Container).RemoveNonFullRoots` method. The method removes
+non-full roots of a feed from database. The method used by node to clean up
+during shutdown
+
 
 ####  4:15 22 June 2017 UTC